# packages/shared-types/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OrganisationOut": {
        "properties": {
          "abn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Abn"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "name",
          "abn",
          "status",
          "created_at",
          "roles",
          "permissions"
        ],
        "title": "OrganisationOut",
        "type": "object"
      },
      "OrganisationUpdate": {
        "properties": {
          "abn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Abn"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "OrganisationUpdate",
        "type": "object"
      },
...
      "Vertical": {
        "enum": [
          "PLANNING",
          "VESSEL",
          "BUSINESS",
          "GRANT",
          "SELL",
          "RENT"
        ],
        "title": "Vertical",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "ApprovalReady API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/organisations/{organisation_id}": {
      "get": {
        "operationId": "get_organisation_v1_organisations__organisation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "organisation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Organisation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Organisation",
        "tags": [
          "organisations"
        ]
      },
      "patch": {
        "operationId": "update_organisation_v1_organisations__organisation_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "organisation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Organisation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganisationUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Organisation",
        "tags": [
          "organisations"
        ]
      }
    },
...
    "/v1/organisations/{organisation_id}/projects": {
      "get": {
        "operationId": "list_projects_v1_organisations__organisation_id__projects_get",
        "parameters": [
          {
            "in": "path",
            "name": "organisation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Organisation Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "vertical",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Vertical"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Vertical"
            }
          },
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectOut"
                  },
                  "title": "Response List Projects V1 Organisations  Organisation Id  Projects Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Projects",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "operationId": "create_project_v1_organisations__organisation_id__projects_post",
        "parameters": [
          {
            "in": "path",
            "name": "organisation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Organisation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Project",
        "tags": [
          "projects"
        ]
      }
    },
    "/v1/organisations/{organisation_id}/projects/{project_id}": {
      "delete": {
        "operationId": "delete_project_v1_organisations__organisation_id__projects__project_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "organisation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Organisation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Project",
        "tags": [
          "projects"
        ]
      },
      "get": {
        "operationId": "get_project_v1_organisations__organisation_id__projects__project_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "organisation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Organisation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Project",
        "tags": [
          "projects"
        ]
      },
      "patch": {
        "operationId": "update_project_v1_organisations__organisation_id__projects__project_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "organisation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Organisation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Project",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/v1/organisations/{organisation_id}/projects/{project_id}/reminders": {
      "get": {
        "operationId": "list_reminders_v1_organisations__organisation_id__projects__project_id__reminders_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "organisation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Organisation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReminderOut"
                  },
                  "title": "Response List Reminders V1 Organisations  Organisation Id  Projects  Project Id  Reminders Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Reminders",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "operationId": "create_reminder_v1_organisations__organisation_id__projects__project_id__reminders_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "organisation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Organisation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReminderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Reminder",
        "tags": [
          "projects"
        ]
      }
    },
    "/v1/organisations/{organisation_id}/projects/{project_id}/reminders/{reminder_id}": {
      "delete": {
        "operationId": "cancel_reminder_v1_organisations__organisation_id__projects__project_id__reminders__reminder_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "reminder_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Reminder Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "organisation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Organisation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Reminder",
        "tags": [
          "projects"
        ]
      }
    },
    "/v1/organisations/{organisation_id}/projects/{project_id}/status": {
      "post": {
        "operationId": "change_project_status_v1_organisations__organisation_id__projects__project_id__status_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "organisation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Organisation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectStatusChange"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Change Project Status",
        "tags": [
          "projects"
        ]
      }
    },
    "/v1/organisations/{organisation_id}/projects/{project_id}/status-events": {
      "get": {
        "operationId": "list_status_events_v1_organisations__organisation_id__projects__project_id__status_events_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "organisation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Organisation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StatusEventOut"
                  },
                  "title": "Response List Status Events V1 Organisations  Organisation Id  Projects  Project Id  Status Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Status Events",
        "tags": [
          "projects"
        ]
      }
    },
    "/v1/organisations/{organisation_id}/projects/{project_id}/submissions": {
      "post": {
        "description": "Start the questionnaire for this project, or return the one already in progress.",
        "operationId": "start_submission_v1_organisations__organisation_id__projects__project_id__submissions_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "organisation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Organisation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmissionStart"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionOut"
                }
              }
            },
            "description": "Created"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Submission",
        "tags": [
          "projects"
        ]
      }
    },
...
  }
}